+++ packages/common/qds-core/src/radio/radio.types.ts
@@ -43,8 +43,16 @@ export interface QdsRadioItemHiddenInputBindings {
   className: RadioClasses["itemHiddenInput"]
 }
 
-export interface QdsRadioErrorTextBindings {
+export interface QdsRadioGroupErrorTextBindings {
   className: InputClasses["errorText"]
+}
+
+export interface QdsRadioGroupHintBindings {
+  className: InputClasses["hint"]
+}
+
+export interface QdsRadioItemHintBindings {
+  className: InputClasses["hint"]
 }
 
 export interface QdsRadioGroupBindings {
@@ -65,12 +73,14 @@ export interface QdsRadioApi {
   size: QdsRadioSize
 
   // group: bindings
-  getErrorTextBindings(): QdsRadioErrorTextBindings
   getGroupBindings(): QdsRadioGroupBindings
+  getGroupErrorTextBindings(): QdsRadioGroupErrorTextBindings
+  getGroupHintBindings(): QdsRadioGroupHintBindings
   getGroupItemsBindings(): QdsRadioGroupItemsBindings
   getGroupLabelBindings(): QdsRadioGroupLabelBindings
   getItemBindings(): QdsRadioItemBindings
   getItemControlBindings(): QdsRadioItemControlBindings
   getItemHiddenInputBindings(): QdsRadioItemHiddenInputBindings
+  getItemHintBindings(): QdsRadioItemHintBindings
   getItemLabelBindings(): QdsRadioItemLabelBindings
 }