+++ packages/docs/react-docs/src/routes/components+/checkbox+/demos/checkbox-explorer-demo.tsx
@@ -3,5 +3,5 @@ import type {ReactElement} from "react"
 import {Checkbox} from "@qualcomm-ui/react/checkbox"
 
 export function CheckboxExplorerDemo(): ReactElement {
-  return <Checkbox checked label="Example Checkbox" />
+  return <Checkbox checked hint="More context here" label="Example Checkbox" />
 }